Integrate FufillmentContext into InferCtxt's transactions
This does not sucessfully build because we incur a double borrow when trying to build libcore, it seems that we
need to finally bite the bullet and make FulfillmentContext completely internal to the InferCtxt. It feels like
it is necessary to finally rebrand the InferCtxt since it has become the typing context.